backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]
sqlalchemy>=1.4.0
psycopg2-binary
alembic
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
aiohttp==3.8.3
web3>=6.0.0,<7.0.0
tensorflow-cpu>=2.12.0
scikit-learn>=1.0.0
numpy>=1.23.5
javalang==0.13.0
solidity-parser==0.0.7
cytoolz==0.12.0
eth-brownie==1.19.3
slither-analyzer>=0.9.0
mythril>=0.23.0
